[refactor] Analog interface AST
