+++ src/oqd_core/interface/analog/__init__.py
@@ -12,156 +12,112 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from oqd_core.interface.analog.operation import (
-    AnalogCircuit,
-    Evolve,
-    Measure,
-    Initialize,
-    Declaration,
-    MyList,
-    QuantumBit,
-    QuantumRegister,
-    Extract,
-    IfElse,
-    While,
-    ModeBit,
-    ModeRegister,
-    Break, 
-    Continue,
-    AnalogExprSubtypes,
-)
-from oqd_core.interface.analog.operator import (
-    Annihilation,
-    Creation,
-    Identity,
-    Ladder,
-    Operator,
-    OperatorAdd,
-    OperatorBinaryOp,
-    OperatorKron,
-    OperatorMul,
-    OperatorScalarMul,
-    OperatorSub,
-    OperatorTerminal,
-    OperatorSubtypes,
-    CastOperator,
-    Pauli,
+from oqd_core.interface.analog.atom import (
+    Access,
+    Atom,
+    Bool,
+    MathImag,
+    MathNum,
+    MathVar,
     PauliI,
-    PauliMinus,
-    PauliPlus,
     PauliX,
     PauliY,
     PauliZ,
 )
+from oqd_core.interface.analog.circuit import AnalogCircuit
 from oqd_core.interface.analog.expression import (
-    Expr,
-    Identifier,
-    Access,
-)
-from oqd_core.interface.analog.math import (
-    MathExpr,
-    MathTerminal,
-    MathStr,
-    MathNum,
-    MathVar,
-    MathImag,
-    MathFunc,
-    MathBinaryOp,
-    MathAdd,
-    MathSub,
-    MathMul,
-    MathDiv,
-    MathPow,
-    MathExprSubtypes,
-    ConstantMathExpr,
-    CastMathExpr,
-)
-from oqd_core.interface.analog.bool import (
+    AnalogExpr,
+    AnalogExprSubtypes,
+    AnalogList,
+    AnalogListExtract,
     BoolAnd,
-    BoolOr,
-    BoolNot,
     BoolEq,
-    BoolNotEq,
-    BoolLessThan,
-    BoolLessThanEq,
+    BoolExpr,
     BoolGreaterThan,
     BoolGreaterThanEq,
-    BoolTrue,
-    BoolFalse,
-    BoolExprSubtypes,
-    BoolExpr,
-    CastBool,
+    BoolLessThan,
+    BoolLessThanEq,
+    BoolNot,
+    BoolNotEq,
+    BoolOr,
+    MathAdd,
+    MathBinaryOp,
+    MathDiv,
+    MathExpr,
+    MathFunc,
+    MathMul,
+    MathPow,
+    MathSub,
+    ModeRegister,
+    OperatorAdd,
+    OperatorKron,
+    OperatorMul,
+    OperatorSub,
+    QuantumBit,
+    QuantumMode,
+    QuantumRegister,
+)
+from oqd_core.interface.analog.statement import (
+    Break,
+    Continue,
+    Declaration,
+    Evolve,
+    IfElse,
+    Initialize,
+    Measure,
+    While,
 )
 
+########################################################################################
+
 __all__ = [
-    "Expr",
-    "Operator",
-    "OperatorTerminal",
-    "OperatorSubtypes",
-    "CastOperator",
-    "Pauli",
+    "Access",
+    "Atom",
+    "Bool",
+    "MathImag",
+    "MathNum",
+    "MathVar",
     "PauliI",
     "PauliX",
     "PauliY",
     "PauliZ",
-    "PauliPlus",
-    "PauliMinus",
-    "Ladder",
-    "Creation",
-    "Annihilation",
-    "Identity",
-    "OperatorBinaryOp",
-    "OperatorAdd",
-    "OperatorSub",
-    "OperatorMul",
-    "OperatorScalarMul",
-    "OperatorKron",
     "AnalogCircuit",
+    "AnalogExpr",
     "AnalogExprSubtypes",
-    "Evolve",
-    "Measure",
-    "Initialize",
-    "Declaration",
-    "MyList",
-    "QuantumBit",
-    "QuantumRegister",
-    "Access",
-    "Extract",
-    "Identifier",
-    "IfElse",
-    "While",
-    "ModeBit",
-    "ModeRegister",
-    "Break", 
-    "Continue",
-    "MathExpr",
-    "MathTerminal",
-    "MathStr",
-    "MathNum",
-    "MathVar",
-    "MathImag",
-    "MathFunc",
-    "MathBinaryOp",
-    "MathAdd",
-    "MathSub",
-    "MathMul",
-    "MathDiv",
-    "MathPow",
-    "MathExprSubtypes",
-    "ConstantMathExpr",
-    "CastMathExpr",
+    "AnalogList",
+    "AnalogListExtract",
     "BoolAnd",
-    "BoolOr",
-    "BoolNot",
     "BoolEq",
-    "BoolNotEq",
-    "BoolLessThan",
-    "BoolLessThanEq",
+    "BoolExpr",
     "BoolGreaterThan",
     "BoolGreaterThanEq",
-    "BoolTrue",
-    "BoolFalse",
-    "BoolExprSubtypes",
-    "BoolExpr",
-    "CastBool",
+    "BoolLessThan",
+    "BoolLessThanEq",
+    "BoolNot",
+    "BoolNotEq",
+    "BoolOr",
+    "MathAdd",
+    "MathBinaryOp",
+    "MathDiv",
+    "MathExpr",
+    "MathFunc",
+    "MathMul",
+    "MathPow",
+    "MathSub",
+    "ModeRegister",
+    "OperatorAdd",
+    "OperatorKron",
+    "OperatorMul",
+    "OperatorSub",
+    "QuantumBit",
+    "QuantumMode",
+    "QuantumRegister",
+    "Break",
+    "Continue",
+    "Declaration",
+    "Evolve",
+    "IfElse",
+    "Initialize",
+    "Measure",
+    "While",
 ]
