--- src/oqd_core/interface/analog/operator.py
@@ -1,320 +1,0 @@
-# Copyright 2024-2025 Open Quantum Design
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-from typing import Any, Union, Annotated
-from pydantic import BeforeValidator, Discriminator, Tag
-
-########################################################################################
-from .math import (
-    CastMathExpr,
-    MathExpr,
-    MathImag,
-    MathMul,
-    MathNum,
-)
-
-from .expression import Expr, Access
-
-########################################################################################
-
-__all__ = [
-    "Operator",
-    "OperatorTerminal",
-    "Pauli",
-    "PauliI",
-    "PauliX",
-    "PauliY",
-    "PauliZ",
-    "PauliPlus",
-    "PauliMinus",
-    "Ladder",
-    "Creation",
-    "Annihilation",
-    "Identity",
-    "OperatorBinaryOp",
-    "OperatorAdd",
-    "OperatorSub",
-    "OperatorMul",
-    "OperatorScalarMul",
-    "OperatorKron",
-    "OperatorSubtypes",
-    "CastOperator",
-]
-
-
-########################################################################################
-
-
-class Operator(Expr):
-    """
-    Class representing the abstract syntax tree (AST) for a quantum operator
-    """
-    
-    @classmethod
-    def cast(cls, value: Any):
-        if isinstance(value, dict):
-            return value
-        if isinstance(value, Operator):
-            return value
-        if isinstance(value, Access):
-            return value
-        raise TypeError
-
-    def __neg__(self):
-        return OperatorScalarMul(op=self, expr=MathNum(value=-1))
-
-    def __pos__(self):
-        return self
-
-    def __add__(self, other):
-        return OperatorAdd(op1=self, op2=other)
-
-    def __sub__(self, other):
-        return OperatorSub(op1=self, op2=other)
-
-    def __matmul__(self, other):
-        if isinstance(other, Operator):
-            return OperatorKron(op1=self, op2=other)
-
-        raise TypeError(
-            f"Attempted Kron product between Operator and {type(other)}. Double check brackets."
-        )
-
-    def __mul__(self, other):
-        if isinstance(other, Operator):
-            return OperatorMul(op1=self, op2=other)
-        else:
-            other = MathExpr.cast(other)
-            return OperatorScalarMul(op=self, expr=other)
-
-    def __rmul__(self, other):
-        other = MathExpr.cast(other)
-        return self * other
-
-    def __truediv__(self, other):
-        other = MathExpr.cast(other)
-        return OperatorScalarMul(op=self, expr=1 / other)
-
-CastOperator = Annotated[Expr, BeforeValidator(Operator.cast)]
-
-########################################################################################
-
-
-class OperatorTerminal(Operator):
-    """
-    Class representing a terminal in the [`Operator`][oqd_core.interface.analog.operator.Operator] abstract syntax tree (AST)
-    """
-
-    pass
-
-
-########################################################################################
-
-
-class Pauli(OperatorTerminal):
-    """
-    Class representing a Pauli operator
-    """
-
-    pass
-
-
-class PauliI(Pauli):
-    """
-    Class for the Pauli I operator
-    """
-
-    pass
-
-
-class PauliX(Pauli):
-    """
-    Class for the Pauli X operator
-    """
-
-    pass
-
-
-class PauliY(Pauli):
-    """
-    Class for the Pauli Y operator
-    """
-
-    pass
-
-
-class PauliZ(Pauli):
-    """
-    Class for the Pauli Z operator
-    """
-
-    pass
-
-
-def PauliPlus():
-    """
-    Function that constructs the Pauli + operator
-    """
-    return OperatorAdd(
-        op1=OperatorScalarMul(op=PauliX(), expr=MathNum(value=0.5)),
-        op2=OperatorScalarMul(
-            op=PauliY(), expr=MathMul(expr1=MathImag(), expr2=MathNum(value=0.5))
-        ),
-    )
-
-
-def PauliMinus():
-    """
-    Function that constructs the Pauli - operator
-    """
-    return OperatorAdd(
-        op1=OperatorScalarMul(op=PauliX(), expr=MathNum(value=0.5)),
-        op2=OperatorScalarMul(
-            op=PauliY(), expr=MathMul(expr1=MathImag(), expr2=MathNum(value=-0.5))
-        ),
-    )
-
-
-########################################################################################
-
-
-class Ladder(OperatorTerminal):
-    """
-    Class representing a ladder operator in Fock space
-    """
-
-    pass
-
-
-class Creation(Ladder):
-    """
-    Class for the Creation operator in Fock space
-    """
-
-    pass
-
-
-class Annihilation(Ladder):
-    """
-    Class for the Annihilation operator in Fock space
-    """
-
-    pass
-
-
-class Identity(Ladder):
-    """
-    Class for the Identity operator in Fock space
-    """
-
-    pass
-
-
-########################################################################################
-
-
-class OperatorScalarMul(Operator):
-    """
-    Class representing scalar multiplication of an [`Operator`][oqd_core.interface.analog.operator.Operator] and a
-    [`MathExpr`][oqd_core.interface.math.MathExpr]
-
-    Attributes:
-        op (Operator): [`Operator`][oqd_core.interface.analog.operator.Operator] to multiply
-        expr (MathExpr): [`MathExpr`][oqd_core.interface.math.MathExpr] to multiply by
-    """
-
-    op: CastOperator
-    expr: CastMathExpr
-
-
-class OperatorBinaryOp(Operator):
-    """
-    Class representing binary operations on [`Operators`][oqd_core.interface.analog.operator.Operator]
-    """
-
-    pass
-
-
-class OperatorAdd(OperatorBinaryOp):
-    """
-    Class representing the addition of [`Operators`][oqd_core.interface.analog.operator.Operator]
-
-    Attributes:
-        op1 (Operator): Left hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-        op2 (Operator): Right hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-    """
-
-    op1: CastOperator
-    op2: CastOperator
-
-
-class OperatorSub(OperatorBinaryOp):
-    """
-    Class representing the subtraction of [`Operators`][oqd_core.interface.analog.operator.Operator]
-
-    Attributes:
-        op1 (Operator): Left hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-        op2 (Operator): Right hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-    """
-
-    op1: CastOperator
-    op2: CastOperator
-
-
-class OperatorMul(OperatorBinaryOp):
-    """
-    Class representing the multiplication of [`Operators`][oqd_core.interface.analog.operator.Operator]
-
-    Attributes:
-        op1 (Operator): Left hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-        op2 (Operator): Right hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-    """
-
-    op1: CastOperator
-    op2: CastOperator
-
-
-class OperatorKron(OperatorBinaryOp):
-    """
-    Class representing the tensor product of [`Operators`][oqd_core.interface.analog.operator.Operator]
-
-    Attributes:
-        op1 (Operator): Left hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-        op2 (Operator): Right hand side [`Operator`][oqd_core.interface.analog.operator.Operator]
-    """
-
-    op1: CastOperator
-    op2: CastOperator
-
-OperatorSubtypes = Annotated[
-    Union[
-        Annotated[PauliI, Tag("PauliI")],
-        Annotated[PauliX, Tag("PauliX")],
-        Annotated[PauliY, Tag("PauliY")],
-        Annotated[PauliZ, Tag("PauliZ")],
-        Annotated[Creation, Tag("Creation")],
-        Annotated[Annihilation, Tag("Annihilation")],
-        Annotated[Identity, Tag("Identity")],
-        Annotated[OperatorScalarMul, Tag("OperatorScalarMul")],
-        Annotated[OperatorAdd, Tag("OperatorAdd")],
-        Annotated[OperatorSub, Tag("OperatorSub")],
-        Annotated[OperatorMul, Tag("OperatorMul")],
-        Annotated[OperatorKron, Tag("OperatorKron")],
-    ],
-    Discriminator(
-        lambda v: v["class_"] if isinstance(v, dict) else getattr(v, "class_")
-    ),
-]
