+++ src/oqd_core/interface/analog/statement.py
@@ -13,65 +13,32 @@
 # limitations under the License.
 
 from __future__ import annotations
+
 from typing import List, Union
 
 from oqd_compiler_infrastructure import TypeReflectBaseModel
-from pydantic.types import NonNegativeInt
 
-from .expression import Expr, Access, Identifier
-from .bool import BoolExprSubtypes, CastBool
-from .math import CastMathExpr, MathExprSubtypes
-from .operator import OperatorSubtypes, CastOperator
+from .atom import Identifier
+from .expression import AnalogExprSubtypes
 
 ########################################################################################
 __all__ = [
-    "AnalogCircuit",
     "Evolve",
     "Measure",
     "Initialize",
-    "QuantumBit",
-    "QuantumRegister",
     "Declaration",
-    "MyList",
-    "Extract",
     "IfElse",
     "While",
-    "ModeBit",
-    "ModeRegister",
-    "Break", 
+    "Break",
     "Continue",
-    "AnalogExprSubtypes"
 ]
 
 ########################################################################################
 
-class MyList(Expr):
-    values: List[AnalogExprSubtypes]
 
 class Declaration(TypeReflectBaseModel):
     name: Identifier
     value: AnalogExprSubtypes
-
-class QuantumBit(TypeReflectBaseModel):
-    access: Access
-    index: NonNegativeInt
-
-
-class QuantumRegister(TypeReflectBaseModel):
-    size: NonNegativeInt
-
-class ModeBit(TypeReflectBaseModel):
-    access: Access
-    index: NonNegativeInt
-
-
-class ModeRegister(TypeReflectBaseModel):
-    size: NonNegativeInt
-
-
-class Extract(TypeReflectBaseModel):
-    access: Access
-    index: NonNegativeInt
 
 
 class Evolve(TypeReflectBaseModel):
@@ -84,8 +51,8 @@ class Evolve(TypeReflectBaseModel):
         targets (Expr): Indices and Quantum objects on which to apply the Hamiltonian
     """
 
-    hamiltonian: CastOperator
-    duration: CastMathExpr
+    hamiltonian: AnalogExprSubtypes
+    duration: AnalogExprSubtypes
     targets: AnalogExprSubtypes
 
 
@@ -93,6 +60,7 @@ class Measure(TypeReflectBaseModel):
     """
     Class representing a measurement in the analog circuit
     """
+
     targets: AnalogExprSubtypes
 
 
@@ -100,6 +68,7 @@ class Initialize(TypeReflectBaseModel):
     """
     Class representing a initialization in the analog circuit
     """
+
     targets: AnalogExprSubtypes
 
 
@@ -107,16 +76,18 @@ class IfElse(TypeReflectBaseModel):
     """
     Class representing a conditional branch in the analog circuit
     """
-    condition: CastBool
+
+    condition: AnalogExprSubtypes
     then_branch: List[Statement] = []
     else_branch: List[Statement] = []
-    
-    
+
+
 class While(TypeReflectBaseModel):
     """
     Class representing a while loop in the analog circuit
     """
-    condition : CastBool
+
+    condition: AnalogExprSubtypes
     body: List[Statement] = []
 
 
@@ -124,6 +95,7 @@ class Break(TypeReflectBaseModel):
     """
     Class representing a break statement to exit the innermost loop
     """
+
     pass
 
 
@@ -131,46 +103,14 @@ class Continue(TypeReflectBaseModel):
     """
     Class representing a continue statement to jump to the next loop iteration
     """
+
     pass
 
 
 """
-Union of classes 
+Union of classes
 """
 
-AnalogExprSubtypes = Union[
-    MathExprSubtypes,
-    OperatorSubtypes,
-    BoolExprSubtypes,
-    MyList,
-    Access,
-    QuantumBit,
-    QuantumRegister,
-    ModeBit,
-    ModeRegister,
-    Extract,
+Statement = Union[
+    Declaration, Measure, Evolve, Initialize, IfElse, While, Break, Continue
 ]
-
-Statement = Union[Declaration, Measure, Evolve, Initialize, IfElse, While, Break, Continue]
-
-
-class AnalogCircuit(TypeReflectBaseModel):
-    """
-    Class representing a quantum information experiment represented in terms of analog operations.
-
-    Attributes:
-        statements (List[Union[Measure, Evolve, Initialize]]): List of statements, including initialize, evolve, measure
-
-    """
-
-    statements: List[Statement] = []
-
-    def evolve(self, hamiltonian: CastOperator, duration: CastMathExpr, targets: AnalogExprSubtypes):
-        self.statements.append(Evolve(hamiltonian=hamiltonian, duration=duration, targets=targets))
-
-    def initialize(self, targets: AnalogExprSubtypes):
-        self.statements.append(Initialize(targets=targets))
-
-    def measure(self, targets: AnalogExprSubtypes):
-        self.statements.append(Measure(targets=targets))
-
